share function from maptooltip
