+++ web/src/components/tooltips/MapTooltip.tsx
@@ -14,37 +14,13 @@ import {
   selectedDatetimeIndexAtom,
   timeAverageAtom,
 } from 'utils/state/atoms';
+import { getTooltipPosition } from './utilities';
 
 interface MapTooltipProperties {
   mousePositionX: number;
   mousePositionY: number;
   hoveredFeature: { featureId: string | number | undefined; zoneId: string };
 }
-
-const getTooltipPosition = (
-  mousePositionX: number,
-  mousePositionY: number,
-  screenWidth: number,
-  tooltipWidth: number,
-  tooltipHeight: number
-) => {
-  const ToolTipFlipBoundaryX = tooltipWidth + 30;
-  const ToolTipFlipBoundaryY = tooltipHeight - 40;
-  const xOffset = 10;
-  const yOffset = tooltipHeight - 40;
-
-  const tooltipPosition = {
-    x: mousePositionX + xOffset,
-    y: mousePositionY - yOffset,
-  };
-  if (screenWidth - mousePositionX < ToolTipFlipBoundaryX) {
-    tooltipPosition.x = mousePositionX - tooltipWidth;
-  }
-  if (mousePositionY < ToolTipFlipBoundaryY) {
-    tooltipPosition.y = mousePositionY;
-  }
-  return tooltipPosition;
-};
 
 function TooltipInner({
   zoneData,
