+++ web/src/components/tooltips/utilities.ts
@@ -1,0 +1,24 @@
+export const getTooltipPosition = (
+  mousePositionX: number,
+  mousePositionY: number,
+  screenWidth: number,
+  tooltipWidth: number,
+  tooltipHeight: number
+) => {
+  const ToolTipFlipBoundaryX = tooltipWidth + 30;
+  const ToolTipFlipBoundaryY = tooltipHeight - 40;
+  const xOffset = 10;
+  const yOffset = tooltipHeight - 40;
+
+  const tooltipPosition = {
+    x: mousePositionX + xOffset,
+    y: mousePositionY - yOffset,
+  };
+  if (screenWidth - mousePositionX < ToolTipFlipBoundaryX) {
+    tooltipPosition.x = mousePositionX - tooltipWidth;
+  }
+  if (mousePositionY < ToolTipFlipBoundaryY) {
+    tooltipPosition.y = mousePositionY;
+  }
+  return tooltipPosition;
+};
